Switch XSrc::Imm to use 32-bit immediates
Previously, XSrc::Imm in the AMD64 codegen would use a 64-bit integer
for its immediate. However, there is generally only one kind of
instruction that can accept a 64-bit immediate: a mov into a register.
Having XSrc::Imm use 64-bit immediates was resulting in a lot of 32-bit
bounds checks being required and only benefitted this one particular
instruction, so it has been removed. Instead, when required, a mov of a
64-bit immediate into a register will simply be added as a special case
instruction.

+++ libhootc/src/codegen/amd64/gen.rs
@@ -16,7 +16,7 @@ struct CodeGenContext {
 
 fn generate_src_for_operand(o: &IlOperand, _ctx: &mut CodeGenContext, _span: IlSpanId) -> XSrc {
     match *o {
-        IlOperand::Const(IlConst::I32(val)) => XSrc::Imm(val as i64),
+        IlOperand::Const(IlConst::I32(val)) => XSrc::Imm(val),
         IlOperand::Register(reg) => XSrc::Reg(SrcRegister::virt(reg))
     }
 }
@@ -31,7 +31,7 @@ fn generate_load_for_operand(o: &IlOperand, ctx: &mut CodeGenContext, span: IlSp
                 InstructionKind::Mov(
                     RegisterSize::DWord,
                     XDest::Reg(DestRegister::virt(None, reg)),
-                    XSrc::Imm(val as i64)
+                    XSrc::Imm(val)
                 ),
                 span
             ));
